moved optimize plotting
