+++ autofit/non_linear/search/optimize/abstract_optimize.py
@@ -3,6 +3,8 @@ from abc import ABC
 from autoconf import conf
 from autofit.non_linear.search.abstract_search import NonLinearSearch
 from autofit.non_linear.samples import Samples
+from autofit.non_linear.plot.optimize_plotters import OptimizePlotter
+from autofit.non_linear.plot.output import Output
 
 
 class AbstractOptimizer(NonLinearSearch, ABC):
@@ -13,3 +15,29 @@ class AbstractOptimizer(NonLinearSearch, ABC):
     @property
     def samples_cls(self):
         return Samples
+
+    @property
+    def plotter_cls(self):
+        return OptimizePlotter
+
+    def plot_results(self, samples):
+
+        def should_plot(name):
+            return conf.instance["visualize"]["plots_search"][self.__class__.__name__][name]
+
+        plotter = OptimizePlotter(
+            samples=samples,
+            output=Output(path=self.paths.image_path / "search", format="png")
+        )
+
+        if should_plot("contour"):
+            plotter.contour()
+
+        if should_plot("cost_history"):
+            plotter.cost_history()
+
+        if should_plot("trajectories"):
+            plotter.trajectories()
+
+        if should_plot("time_series"):
+            plotter.time_series()
